refactor(Map): Refactor
- Refactored new configuration format for boundary layer
