+++ src/modules/Intervention/components/SubLevelAnalysis/components/Map/components/BoundaryLayer/hooks/config.ts
@@ -1,0 +1,9 @@
+import { useParams } from "react-router-dom";
+import { useRecoilValue } from "recoil";
+import { BoundaryMapLayer } from "../../../../../../../../../shared/interfaces/interventionConfig";
+import { InterventionStateSelector } from "../../../../../../../state/intervention";
+
+export default function useBoundaryConfig(): BoundaryMapLayer {
+  const { id } = useParams();
+  return useRecoilValue(InterventionStateSelector({ id, path: ["map", "coreLayers", "boundaryLayer"] }));
+}
