+++ src/modules/Intervention/components/SubLevelAnalysis/components/Map/state/data.ts
@@ -1,39 +1,11 @@
-import { find, findIndex, flatten, groupBy } from "lodash";
+import { find, findIndex, groupBy } from "lodash";
 import { selectorFamily } from "recoil";
 import { EngineState } from "../../../../../../../core/state/dataEngine";
 import { InterventionOrgUnitState, InterventionPeriodState } from "../../../../../state/selections";
 import { SubLevelOrgUnit } from "../../../state/dimensions";
-import { getAnalyticsData, getBoundaryData } from "../services/data";
-import { convertCoordinates } from "../utils/map";
+import { BoundaryData } from "../components/BoundaryLayer/state/data";
+import { getAnalyticsData } from "../services/data";
 import { MapIndicatorState } from "./config";
-
-export const BoundaryData = selectorFamily<any, string | undefined>({
-  key: "boundary-data",
-  get:
-    (id?: string) =>
-    async ({ get }) => {
-      if (!id) return null;
-      const subLevelOrgUnit = get(SubLevelOrgUnit(id));
-      const selectedOrgUnit = get(InterventionOrgUnitState(id));
-      const engine = get(EngineState);
-      if (!subLevelOrgUnit) return null;
-      const boundaryData = await getBoundaryData(engine, [selectedOrgUnit.id, ...subLevelOrgUnit]);
-      return flatten(
-        boundaryData?.map((area: { co: string; id: string; na: string; le: number }) => ({
-          id: area.id,
-          name: area.na,
-          level: area.le,
-          co: flatten(JSON.parse(area.co)).map((points: any) => {
-            if (!points) return [];
-            if (typeof points[0] === "number") {
-              return convertCoordinates(points);
-            }
-            return points?.map(convertCoordinates);
-          }),
-        }))
-      );
-    },
-});
 
 export const MapIndicatorData = selectorFamily<any, string | undefined>({
   key: "map-indicator-data",
